:recycle: split external_resource.tasks module
