+++ packages/automated_actions/automated_actions/celery/external_resource/_elasticache_flush.py
@@ -1,0 +1,85 @@
+from automated_actions_utils.cluster_connection import get_cluster_connection_data
+from automated_actions_utils.external_resource import (
+    ExternalResource,
+    ExternalResourceProvider,
+    get_external_resource,
+    settings,
+)
+from automated_actions_utils.openshift_client import (
+    OpenshiftClient,
+    SecretKeyRef,
+    job_builder,
+)
+
+from automated_actions.celery.app import app
+from automated_actions.celery.automated_action_task import AutomatedActionTask
+from automated_actions.db.models import Action
+
+
+class ExternalResourceFlushElastiCache:
+    def __init__(
+        self, action: Action, oc: OpenshiftClient, elasticache: ExternalResource
+    ) -> None:
+        self.action = action
+        self.oc = oc
+        self.elasticache = elasticache
+
+    def run(
+        self,
+        image: str,
+        command: list[str],
+        args: list[str],
+        secret_name: str,
+        env_secret_mappings: dict[str, str],
+    ) -> None:
+        job = job_builder(
+            image=image,
+            command=command,
+            args=args,
+            job_name_prefix="flush-elasticache-",
+            annotations={
+                "automated-actions.action_id": str(self.action.action_id),
+            },
+            env_secrets={
+                key: SecretKeyRef(
+                    secret=secret_name,
+                    key=value,
+                )
+                for key, value in env_secret_mappings.items()
+            },
+        )
+        return self.oc.run_job(namespace=self.elasticache.namespace, job=job)
+
+
+@app.task(base=AutomatedActionTask)
+def external_resource_flush_elasticache(
+    account: str,
+    identifier: str,
+    *,
+    action: Action,
+) -> None:
+    elasticache = get_external_resource(
+        account=account,
+        identifier=identifier,
+        provider=ExternalResourceProvider.ELASTICACHE,
+    )
+
+    cluster_connection = get_cluster_connection_data(elasticache.cluster, settings)
+    oc = OpenshiftClient(
+        server_url=cluster_connection.url, token=cluster_connection.token
+    )
+    if not elasticache.output_resource_name:
+        raise ValueError(
+            f"Output resource name not defined for {elasticache.identifier} in {elasticache.namespace} namespace.",
+        )
+    ExternalResourceFlushElastiCache(
+        action=action,
+        oc=oc,
+        elasticache=elasticache,
+    ).run(
+        image=settings.external_resource_elasticache.image,
+        command=settings.external_resource_elasticache.flush_command,
+        args=settings.external_resource_elasticache.flush_command_args,
+        secret_name=elasticache.output_resource_name,
+        env_secret_mappings=settings.external_resource_elasticache.env_secret_mappings,
+    )
